Split domain entities into one file per entity with its repository
Each entity, its methods, and its repository interface are co-located
in a single file. Tests split correspondingly.
