--- services/afl/internal/domain/entities.go
@@ -1,149 +1,0 @@
-package domain
-
-import "time"
-
-// MatchResult represents the outcome of a match.
-type MatchResult string
-
-const (
-	MatchResultHomeWin  MatchResult = "home_win"
-	MatchResultAwayWin  MatchResult = "away_win"
-	MatchResultDraw     MatchResult = "draw"
-	MatchResultNoResult MatchResult = "no_result"
-)
-
-// PremiershipPoints awarded per match result.
-const (
-	PremiershipPointsWin  = 4
-	PremiershipPointsDraw = 2
-	PremiershipPointsLoss = 0
-)
-
-// PointsPerGoal is the number of points a goal is worth in AFL.
-const PointsPerGoal = 6
-
-type League struct {
-	ID   int
-	Name string
-}
-
-type Season struct {
-	ID       int
-	Name     string
-	LeagueID int
-}
-
-type Round struct {
-	ID       int
-	Name     string
-	SeasonID int
-}
-
-type Club struct {
-	ID   int
-	Name string
-}
-
-type Match struct {
-	ID        int
-	RoundID   int
-	Home      ClubMatch
-	Away      ClubMatch
-	Venue     string
-	StartTime time.Time
-	Result    MatchResult
-}
-
-// Winner returns a pointer to the winning ClubMatch, or nil for a draw.
-func (m *Match) Winner() *ClubMatch {
-	homeScore := m.Home.Score()
-	awayScore := m.Away.Score()
-	if homeScore > awayScore {
-		return &m.Home
-	}
-	if awayScore > homeScore {
-		return &m.Away
-	}
-	return nil
-}
-
-type ClubSeason struct {
-	ID                  int
-	ClubID              int
-	SeasonID            int
-	Played              int
-	Won                 int
-	Lost                int
-	Drawn               int
-	For                 int
-	Against             int
-	PremiershipPoints   int
-}
-
-type ClubMatch struct {
-	ID            int
-	MatchID       int
-	ClubSeasonID  int
-	RushedBehinds int
-	StoredScore   int
-	PlayerMatches []PlayerMatch
-}
-
-// Score computes the total score from player contributions and rushed behinds.
-func (cm ClubMatch) Score() int {
-	total := cm.RushedBehinds
-	for _, pm := range cm.PlayerMatches {
-		total += pm.Score()
-	}
-	return total
-}
-
-type Player struct {
-	ID     int
-	Name   string
-	ClubID int
-}
-
-type PlayerSeason struct {
-	ID           int
-	PlayerID     int
-	ClubSeasonID int
-}
-
-type PlayerMatch struct {
-	ID             int
-	ClubMatchID    int
-	PlayerSeasonID int
-	Kicks          int
-	Handballs      int
-	Marks          int
-	Hitouts        int
-	Tackles        int
-	Goals          int
-	Behinds        int
-}
-
-// Disposals returns the total disposals (kicks + handballs).
-func (pm PlayerMatch) Disposals() int {
-	return pm.Kicks + pm.Handballs
-}
-
-// Score returns the player's scoring contribution in points (goals * 6 + behinds).
-func (pm PlayerMatch) Score() int {
-	return pm.Goals*PointsPerGoal + pm.Behinds
-}
-
-
-// UpsertPlayerMatchParams holds optional fields for creating or updating a PlayerMatch.
-// Nil fields are left unchanged on update.
-type UpsertPlayerMatchParams struct {
-	ClubMatchID    int
-	PlayerSeasonID int
-	Kicks          *int
-	Handballs      *int
-	Marks          *int
-	Hitouts        *int
-	Tackles        *int
-	Goals          *int
-	Behinds        *int
-}
