+++ services/afl/internal/domain/league.go
@@ -1,0 +1,6 @@
+package domain
+
+type League struct {
+	ID   int
+	Name string
+}
